* 64bit fixup.
git-archimport-id: falk@cs.fau.de--2008-systemoc-intern/SysteMoC-Testcases--mainline--0.2--patch-5

+++ systemoc-wsdf-recons-dila/CImg.h
@@ -4173,7 +4173,7 @@ namespace cimg_library {
 
     //! Set the window title
     CImgDisplay& set_title(const char *format,...) {
-      fps_timer = 0*(int)format;
+      fps_timer = 0*(long)format;
       return *this; 
     }
